comment cluster controller, move it all in the same file
